Makefile.am: use built-in pkginclude_HEADERS variable.
The existing Makefile.am did,

  headersdir = $(includedir)/planarity/
  headers_HEADERS = ...

with the goal of installing all of the headers_HEADERS into a
"planarity" subdirectory of whatever directory the headers would
normally be installed into. This feature is already built-in to
automake, however:

  https://www.gnu.org/software/automake/manual/automake.html#Uniform

Thus the two lines above can be replaced by merely

  pkginclude_HEADERS = ...

which is what this commit does.