autotools: distribute and install samples to $(docdir)/samples.
The sample files located in c/samples are not currently distributed to
end users, because autotools doesn't know about them. This commit
changes that:

  1. First, we add a Makefile.am to the c/samples directory that
     lists all of the documentation files to be installed, and
     specifies where (in a "samples" subdirectory) to install them.

  2. Next, we add a "SUBDIRS" line to the top-level Makefile.am,
     to let it know that there's another one in c/samples.

  3. Finally, we tell configure.ac that it should process the new
     Makefile.am, because ./configure should substitute the value
     of the user's --docdir option into it as $(docdir).

The end result of this is that after running, say,

  $ ./configure --docdir=/usr/local/share/doc/planarity
  $ make
  $ make install

the samples are installed to /usr/local/share/doc/planarity/samples.
This will generally be the default behavior when --docdir takes its
default value, but the user now also has the ability to override it.