Issue 69 - Ensure edge returned by `_getNextEdge()` has endpoints in the upper-triangular order (#70)
Closes #69 

The bug that is fixed here affected the behavior of `RandomGraph()` in `planarityRandomGraphs.c` when the `gp_write()` was changed from writing adjacency list to writing G6. The bug would also have affected future plans to enable `RandomGraphs()` to write the random graphs generated into a single G6 file (which would be easier to manage than a directory of many adjacency list files). 

When an edge is added , it results in a pair of twin arcs in the edge array `E`. When reading adjacency list files with `gp_read()`, the twin arcs are always generated such that the first arc has the lower numbered neighbor value and the second arc has the higher numbered neighbor value.  This same behavior is not guaranteed when `gp_CreateRandomGraph()` adds edges. This fix to the G6 writer ensures that pairs of arcs in the edge array are always interpreted as lesser vertex first then greater so that the vertices give the coordinates in an _upper triangular_ matrix.