Changing quoted string style so that string literals are enclosed in single quotes, but error messages and formatted strings are enclosed in double-quotes
According to PEP-8:
"In Python, single-quoted strings and double-quoted strings are the same. This PEP does not make a recommendation for this. Pick a rule and stick to it. When a string contains single or double quote characters, however, use the other one to avoid backslashes in the string. It improves readability."