Adding --enable-compile-warnings to compile-warnings.yml job step to run ./configure (#156)
Confirmed that I'm able to run the steps in compile-warnings.yml on my Windows box in MSYS2 UCRT terminal, i.e.

autoreconf -fi
./configure --enable-compile-warnings CC="$CC" CFLAGS="$CFLAGS"
make -j4

This, along with all the work by Joseph Edwards, closes #152. 

It's better to have this committed now and then continue to fail the warnings checks until all the warnings are fixed (Issue #147) and all of the sanitizer checks are addressed (Issue #154).